Added support for ISR analysis triggers.
Added bit enums and const bit masks for the following chains:
  EF_e18vh_medium1_2e7T_medium1
  EF_3mu6
  EF_mu18_tight_2mu4_EFFS
  EF_2e7T_medium1_mu6
  EF_e7T_medium1_2mu6
  EF_xe80_tclcw_loose
  EF_j110_a4tchad_xe90_tclcw_loose
  EF_j80_a4tchad_xe100_tclcw_loose
  EF_j80_a4tchad_xe70_tclcw_dphi2j45xe10

We ran out of 32 bits, however, so I needed to update the definitions from int to long long.
"long" would be sufficient for 64 bit releases, but 32 bit releases are still prevalent
and a long int is only 32 bits. So, all of the TRIG_* masks are now long long, and the
trigFlags members for the Event class, Lepton class, and Tau class are all now long long.
Hopefully the TStreamer is able to handle old ROOT files where the variables are saved as
unsigned int and properly cast them into long long. I do not know for sure if that will work.