Adding MCProduction enum to SusyDefs.
This enum will allow to specify which MC production campaign a sample was made in.
This is important when initializing SUSYTools, because some settings (such as JES)
actually depend on the production campaign.

I assume the need for this specification is *temporary*, as this is the first
time people are using a mix of two productions simultaneously.