MCWeighter::getMCWeight: add protection against sumw=0
for the pmssm samples, I am seeing events that get a weight 'nan'.
This is because when computing the per-process sumw, there are a
handful of events with processes for which the 'genCutFlow___'
histogram has no entries in the bin 'SusyProp Veto'.
Is this always the bin to look up?
I don't know whether this might be a bug upstream: these are pmssm
samples, and I thought that the SusyPropagator issue was happening
only for simplified models. More details in the pull request.

This should be investigated and fixed, but at least for now we can
process these samples (and make a cutflow and compute limits).