allow the user to specify the bin used to compute sumw
This change is to deal with the pmssm samples that have incorrect
counts in the "SusyProp" bin in `n015*` ntuples. While that problem
will be fixed upstream, this solution allows the user to get the
correct normalization by calling for the samples in question:

```
m_mcWeighter->setLabelBinCounter("Initial").clearAndRebuildSumwMap(m_tree);
```

This avoids having hard-coded special cases in MCWeighter (they don't
belong here). Details of the implementation:
- `buildSumwMapFromChain()`, `buildSumwMapFromTree()` now private;
  user should use `buildSumwMap()`
- add convenience `sumwmapHasKey()`
- add `m_labelBinCounter`, `setLabelBinCounter()`,
  `defaultLabelBinCounter()`
- warn on the first N occurrences of sumwmap=0
- clarify `getCrossSection()` logic (avoid multiple return statements)

Tested with Herwigpp_UEEE3_CTEQ6L1_DGnoSL_TB10_050_100_100_2L.