Selector2Lep: enum tauid, define get_* functions
trying to re-implement the selection within a TPySelector.
It turns out I can call all the SusyNtTools functions from
python (see next commit of cutflow_2lep.py), so there's no
need of a TSelector, which just makes things more difficult
to debug. This is just a temporary commit for reference.