add JetSelector, and delegate from SusyNtTools
JetSelector is still a draft but it works fine and the cutflow is
unchanged. These are the functions that have been delegated from
SusyNtTools to JetSelector:
- isSignalJet
- isSignalJet2Lep
- isCentralLightJet
- isCentralBJet
- isForwardJet
I plan to move more configurable parameters to JetSelector, since now
~everything is passed in through the c'tor, as suggested in "Working
Effectively with Legacy Code". Also, the analysis-specific values are
still the ones from SusyDefs (but they will be moved here shortly).