drop SusyNtAna::buildSumwMap and SusyNtTools::buildSumwMap
These two methods are obsolete and should not be used anywhere anymore.
MCWeighter should be used instead to build a sumwmap.

Leaving in SusyNtAna::m_sumwMap (as well as its set/get methods)
since they are probably being used in many places; one can still
set this ptr to a valid map provided by MCWeighter.

This commit was already suggested in #14; #33 seems a good place
to put this in.