unless specified, use the "Initial" bin for all samples except simplified models
Determine from the dsid whether this is a simplified model.
Unless the user has specified a non-default normalization bin,
when calling MCWeighter::buildSumwMapFromTree the correct bin
is picked for std samples and simplified model ones.

The unit test in test_mcWeighter.cxx checks that our logic is at least
self-consistent.