MCWeighter::buildSumwMapFromTree: add special case for proc-independent samples
I don't know whether the sumw=0 is due to the fact that the proc is
not being filled for the new xaod samples, but I cannot track back how
the process-independent case was accounted before. This seems to work,
will probably need to review it when we start using process-dependent
samples with xaod.