Specify splitlevel for TTree::Bronch
Without this change, root6 gives the following core dump
```
Fatal in <TBranchElement::InitializeOffsets>: Could not find the real data member 'electrons' when constructing the branch 'electrons' [Likely an internal error, please report to the developers].
aborting
```
It's not clear whether this is the correct solution (re-running
a newer version of D3pdMakerReader does not apply this change).
But for now that's the only solution we have.
The other parameter `bufsize` is left to its default value.