fix normalization bug sumw = N_raw while it should be N_weighted
I introduced this bug when we switched the default bin label f5618d5
The normalization should be done with the total weight, not just with
the total number of events. This shows up, for example, in the crazy
normalization observed by Danny in the Sherpa Zll n0203 samples.
It was introduced in Jan, so it did not affect any official result.
We need to run a weighted cutflow in our unit tests!