Improve MuonId
Same as what done for ElectronId
- muID -> MuonId
- function to convert to string

Note that MuonId is an `enum class`, while ElectronId is not.
We should move everywhere to either `enum class` or `struct { enum }`
(see for example `hlfv::LeptonTruthType`). Typed enums are safer; the
struct solution is even better (IMO) because it allows to define the
various conversions once (eg. to size_t when using the enum to index
an array).