add script/executable for checkign ST xsec
- To check if the xsec * eff * kfactor as returned from the SUSYTools'
xsec database is non-negative (i.e. not found). This will let you know
if a specific samples' xsec information is availalbe or if there is a
problem with your setup.

- Currently this method calls MCWeighter::getXsecTimesEff directly.
  Shortly, the method "getMCWeight" will cause a job to exit if the
  xsec * eff * kfactor is not found. The scripts commited here will then
allow the user to run over quickly all samples to see which ones are OK
(or not) prior to launching jobs.