set default sumw to 0 from -1
having the default and intiail value set to -1 makes for incorrect sumw
calculation when doing:

total sumw = initail sumw + sumw for file

as it will be off by 1! The idea of -1 was initially as a check that the
sumw has not been set but this I don't think is necessary anymore.